New setting; searcharr_allow_group_interaction
This addition prevents other users in a group chat from interacting with conversation with the bot via the InlineKeyboardButtons.

I believe this should be configured as the default behaviour but it can be set to function as previously implemented by changing the searcharr_allow_group_interaction flag to True in the settings.