Markdown formatting, MessageHandler, and User view
Implemented the use of MarkdownV2 in messages to allow for bolding in message captions. Currently only in a work-in-progress function though. Added an ignore for flake8 compliance as the escape of certain characters is required for this parse_mode but it raises a warning as it's particular to this.

A MessageHandler was added to the dispatcher so that in the future, replying to the bot during certain states will allow the user to provide input for things such as name of a permission or access group being added or removed, and perhaps allow for a better experience if commands to search such as /series can be replied to instead of just a syntax help prompt being returned.

A user_view has been added so that clicking on a user when browsing the list will display permissions they have been granted and access groups they belong to. From there they can be added and removed as well.

Permissions and Access Groups command modules created although they are currently placeholders.